Remove non incremental INotifyPropertyChangedGenerator
