--- CommunityToolkit.Mvvm.SourceGenerators/ComponentModel/INotifyPropertyChangedGenerator.Incremental.cs
@@ -1,78 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System.Collections.Immutable;
-using System.Linq;
-using CommunityToolkit.Mvvm.SourceGenerators.Diagnostics;
-using CommunityToolkit.Mvvm.SourceGenerators.Extensions;
-using CommunityToolkit.Mvvm.SourceGenerators.Input.Models;
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CSharp.Syntax;
-using static CommunityToolkit.Mvvm.SourceGenerators.Diagnostics.DiagnosticDescriptors;
-
-namespace CommunityToolkit.Mvvm.SourceGenerators;
-
-/// <summary>
-/// A source generator for the <c>INotifyPropertyChangedAttribute</c> type.
-/// </summary>
-[Generator(LanguageNames.CSharp)]
-public sealed class INotifyPropertyChangedGenerator2 : TransitiveMembersGenerator2<INotifyPropertyChangedInfo>
-{
-    /// <summary>
-    /// Initializes a new instance of the <see cref="INotifyPropertyChangedGenerator2"/> class.
-    /// </summary>
-    public INotifyPropertyChangedGenerator2()
-        : base("global::CommunityToolkit.Mvvm.ComponentModel.INotifyPropertyChangedAttribute")
-    {
-    }
-
-    /// <inheritdoc/>
-    protected override INotifyPropertyChangedInfo GetInfo(AttributeData attributeData)
-    {
-        if (attributeData.TryGetNamedArgument("IncludeAdditionalHelperMethods", out bool includeAdditionalHelperMethods))
-        {
-            return new(includeAdditionalHelperMethods);
-        }
-
-        return new(false);
-    }
-
-    /// <inheritdoc/>
-    protected override bool ValidateTargetType(
-        INamedTypeSymbol typeSymbol,
-        INotifyPropertyChangedInfo info,
-        out ImmutableArray<Diagnostic> diagnostics)
-    {
-        ImmutableArray<Diagnostic>.Builder builder = ImmutableArray.CreateBuilder<Diagnostic>();
-
-        // Check if the type already implements INotifyPropertyChanged
-        if (typeSymbol.AllInterfaces.Any(i => i.HasFullyQualifiedName("global::System.ComponentModel.INotifyPropertyChanged")))
-        {
-            builder.Add(DuplicateINotifyPropertyChangedInterfaceForINotifyPropertyChangedAttributeError, typeSymbol, typeSymbol);
-
-            diagnostics = builder.ToImmutable();
-
-            return false;
-        }
-
-        diagnostics = builder.ToImmutable();
-
-        return true;
-    }
-
-    /// <inheritdoc/>
-    protected override ImmutableArray<MemberDeclarationSyntax> FilterDeclaredMembers(INotifyPropertyChangedInfo info, ClassDeclarationSyntax classDeclaration)
-    {
-        // If requested, only include the event and the basic methods to raise it, but not the additional helpers
-        if (!info.IncludeAdditionalHelperMethods)
-        {
-            return classDeclaration.Members.Where(
-                static member => member
-                    is EventFieldDeclarationSyntax
-                    or MethodDeclarationSyntax { Identifier.ValueText: "OnPropertyChanged" }).ToImmutableArray();
-        }
-
-        return classDeclaration.Members.ToImmutableArray();
-    }
-}
